와인 샵이 아닌, 지역으로 검색할 수 있도록 기능 추가.

+++ wineFineThankyou/Presentation/Shop/WinePriceCalculator.swift
@@ -33,7 +33,7 @@ class WinePriceCalculator {
         }
     }
     
-    class func getRange(_ price: Int) -> String {
+    class func getRange(_ price: Int64) -> String {
         if price < 10000 {
             return PriceRange.none.str
         } else if 10000 <= price, price < 30000 {
